Code Improvements
